requirements.txt
# Core dependencies
asyncio
websockets>=11.0.0
aiohttp>=3.8.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web3 and blockchain
web3>=6.0.0
eth-account>=0.9.0

# Database
sqlite3  # Built-in to Python
sqlalchemy>=2.0.0

# Monitoring and alerts
schedule>=1.2.0
smtplib  # Built-in to Python

# Dashboard
flask>=2.3.0
flask-socketio>=5.3.0
plotly>=5.15.0

# Logging
coloredlogs>=15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0